+++ lite/api-modules/antstream/native/ant_stream_native.cmake
@@ -1,7 +1,7 @@
 get_filename_component(MODULE_DIR ${CMAKE_CURRENT_LIST_FILE} DIRECTORY)
 
-set(MODULE_NAME "ant")
+set(MODULE_NAME "antstream")
 
 add_subdirectory(${MODULE_DIR}/internal/ ${CMAKE_BINARY_DIR}/out/ant/)
 
-list(APPEND EXTERNAL_LIBS ant_native_stream ant_native_ml)
+list(APPEND EXTERNAL_LIBS ant_stream_native)
