+++ lite/test/build-ant-tx2.sh
@@ -2,8 +2,12 @@
 # Build ANT API and App Runtime for IoT.js
 MODULE_PATH=../../lite/api-modules/
 cd ../../dep/iotjs/
-./tools/build.py --target-board=tx2 --cmake-param=-DENABLE_MODULE_ANT=ON \
-    --external-modules=${MODULE_PATH}/ant
+./tools/build.py --target-board=tx2 \
+    --cmake-param=-DENABLE_MODULE_ANT=ON \
+    --cmake-param=-DENABLE_MODULE_ANTCOMPANION=ON \
+    --cmake-param=-DENABLE_MODULE_ANTML=ON \
+    --cmake-param=-DENABLE_MODULE_ANTSTREAM=ON \
+    --external-modules=${MODULE_PATH}
 cd -
 
 # Install ANT API and App Runtime for IoT.js
